requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.14.0
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
drf-yasg==1.21.6
sqlparse==0.5.3
